requirements/base.txt
Django==1.7.7
Markdown==2.6.1
MySQL-python==1.2.5
Pillow==2.7.0
django-braces==1.4.0
django-filter==0.9.2
django-guardian==1.2.5
django-oauth-toolkit==0.7.2
djangorestframework==3.1.0
oauthlib==0.7.2
redis==2.10.3
redisco==0.1.4
six==1.9.0
wsgiref==0.1.2